# docs/api/openapi.yaml
components:
  schemas:
    handler.listResponse:
      properties:
        notifications:
          items:
            $ref: '#/components/schemas/handler.notificationItem'
          type: array
          uniqueItems: false
        total:
          type: integer
      type: object
    handler.messageDetail:
      properties:
        channel:
          type: string
        confidence_score:
          type: number
        content:
          type: string
        created_at:
          type: string
        id:
          type: string
        importance_score:
          type: number
        message_id:
          type: string
        reasoning:
          type: string
        resolved_at:
          type: string
        sender:
          type: string
        source:
          type: string
        source_account:
          type: string
        status:
          type: string
        updated_at:
          type: string
      type: object
    handler.messageListItem:
      properties:
        channel:
          type: string
        confidence_score:
          type: number
        content:
          type: string
        created_at:
          type: string
        id:
          type: string
        importance_score:
          type: number
        sender:
          type: string
        source:
          type: string
        source_account:
          type: string
        status:
          type: string
      type: object
    handler.messagesListResponse:
      properties:
        messages:
          items:
            $ref: '#/components/schemas/handler.messageListItem'
          type: array
          uniqueItems: false
        total:
          type: integer
      type: object
    handler.notificationItem:
      properties:
        channel:
          type: string
        confidence_score:
          type: number
        content:
          type: string
        created_at:
          type: string
        id:
          type: string
        importance_score:
          type: number
        sender:
          type: string
        source:
          type: string
        source_account:
          type: string
      type: object
    handler.timerResponse:
      properties:
        block_type:
          type: string
        display_time:
          type: string
        duration_seconds:
          type: integer
        elapsed_fraction:
          type: number
        elapsed_seconds:
          type: integer
        remaining_seconds:
          type: integer
        running:
          type: boolean
        task_name:
          type: string
      type: object
externalDocs:
  description: WebSocket event channel reference
  url: https://git.hrafn.xyz/aether/cue/src/branch/main/docs/api/websocket.md
info:
  description: |-
    Cue is a local-first, privacy-centric productivity assistant.
    This REST API exposes messages, notifications, feedback buffer,
    tasks, day planner, routing rules, service configuration, timer,
    auth/pairing, and event replay. The WebSocket event channel at
    /api/v1/websocket/events is documented separately in
    docs/api/websocket.md.
  title: Cue API
  version: "1.0"
openapi: 3.1.0
paths:
  /api/v1/events:
    get:
      description: |-
        Returns all WebSocket events with sequence numbers greater than
        `since`. Clients use this after reconnecting to catch up on
        events missed while disconnected. Payload is a JSON array of
        EventEnvelope objects matching those pushed over the WebSocket
        channel; see docs/api/websocket.md for the envelope schema.
      parameters:
      - description: Last seq received by the client
        in: query
        name: since
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  type: object
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: invalid since parameter
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: internal error
      summary: Replay buffered events
      tags:
      - events
  /api/v1/health:
    get:
      description: |-
        Returns 200 {"status":"ok"} whenever the HTTP server is running.
        Does not inspect any subsystem; use /health/ready for readiness.
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
      summary: Liveness check
      tags:
      - health
  /api/v1/health/ready:
    get:
      description: |-
        Runs every registered subsystem check and returns 200 only if
        all report healthy. Returns 503 with per-subsystem error detail
        when any check fails. The response body always includes a
        "subsystems" object keyed by subsystem name.
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
        "503":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Service Unavailable
      summary: Readiness check
      tags:
      - health
  /api/v1/messages:
    get:
      description: |-
        Paginated list of all ingested messages, filterable by status,
        source (slack/email), channel, and since (RFC3339).
      parameters:
      - description: Notified | Buffered | Ignored | Resolved | Pending
        in: query
        name: status
        schema:
          type: string
      - description: slack | email
        in: query
        name: source
        schema:
          type: string
      - description: Source-native channel / folder
        in: query
        name: channel
        schema:
          type: string
      - description: RFC3339 lower bound on created_at
        in: query
        name: since
        schema:
          type: string
      - description: Page size (default 50)
        in: query
        name: limit
        schema:
          type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.messagesListResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: List messages
      tags:
      - messages
  /api/v1/messages/{id}:
    get:
      parameters:
      - description: Message UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.messageDetail'
          description: OK
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Get message by ID
      tags:
      - messages
  /api/v1/notifications:
    get:
      description: |-
        Paginated list of messages currently in the Notified state,
        sorted by created_at descending.
      parameters:
      - description: Page size (default 50)
        in: query
        name: limit
        schema:
          type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.listResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: List active notifications
      tags:
      - notifications
  /api/v1/notifications/{id}:
    get:
      parameters:
      - description: Notification / message UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.messageDetail'
          description: OK
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Get notification by ID
      tags:
      - notifications
  /api/v1/notifications/{id}/dismiss:
    post:
      description: Marks the notification as Ignored. Unlike resolve, no resolved_at
        timestamp is set.
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.messageDetail'
          description: OK
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Dismiss a notification
      tags:
      - notifications
  /api/v1/notifications/{id}/resolve:
    post:
      description: |-
        Marks the notification as resolved and sets resolved_at to now.
        Returns 409 if the notification is already resolved.
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.messageDetail'
          description: OK
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: already resolved
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Resolve a notification
      tags:
      - notifications
  /api/v1/timer:
    get:
      description: |-
        Returns the live Pomodoro timer state derived from today's schedule.
        When no schedule exists or no block is currently active, responds
        with {"running": false} and 200 OK.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.timerResponse'
          description: OK
      summary: Current timer state
      tags:
      - timer
  /health:
    get:
      description: |-
        Returns 200 {"status":"ok"} whenever the HTTP server is running.
        Does not inspect any subsystem; use /health/ready for readiness.
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
      summary: Liveness check
      tags:
      - health
  /health/ready:
    get:
      description: |-
        Runs every registered subsystem check and returns 200 only if
        all report healthy. Returns 503 with per-subsystem error detail
        when any check fails. The response body always includes a
        "subsystems" object keyed by subsystem name.
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
        "503":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Service Unavailable
      summary: Readiness check
      tags:
      - health